plugins/observatory/skills/observatory/pyproject.toml
[project]
name = "claude-code-observatory"
version = "1.0.0"
description = "Claude Code Observatory - Django REST API and Admin for Claude Code data"
readme = "SKILL.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "django-filter>=24.0",
    "django-cors-headers>=4.3",
    "drf-spectacular>=0.27",
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "python-dotenv>=1.0",
    "whitenoise>=6.6",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.2",
    "ipython>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "observatory",
    "sessions",
    "plugins",
    "skills",
    "commands",
    "agents",
    "hooks",
    "settings",
    "output_styles",
    "mcps",
    "knowledge",
]

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar>=4.2",
    "ipython>=8.0",
]
